pass parameters directly to the role for repeatability
Instead of using global vars set in the playbook, use vars passed to the
role. This lets us include the role multiple times with different schema
files to produce different output files.

Fix a bug in the template that shows up when there is an option defined
to control dynamic groups but the option is not set.

Signed-off-by: Doug Hellmann <doug@doughellmann.com>